feat: POC for generically verifying incoming data
When a request is made to issue a credential the signature of the data
being received must be verified to ensure the validity of the data.
The verification should check that the data
 - Was generated by us
 - Is unmodified
 - Has not expired

This POC aims to achieve two goals
 1. The signature verification should be handled in a single place,
    rather then relying on each endpoint to perform the checks.
 2. The verification should not be tightly coupled to the stucture of
    the `tis-trainee-details` DTO, i.e. a new field in the
    `tis-trainee-details` DTO requiring a new field adding to the
    credentials service DTO too.

The POC uses a generic `SignedDataFilter` to achieve these goals, the
filter is applied automatically to any `/api/issue/*` endpoint and by
verifying the raw JSON it does not care about the DTO structure.
The Placement/Programme DTOs used in the IssueResource endpoints can
then only include fields relevant to the credential. All other fields
will be discarded, which prevents any tight coupling.

There is a limitation when using filters in that, by default, the
request body is InputStream based and can only be read a single time. To
avoid breaking any downstream filters, a `CachedBody` input stream and
request have been created. The request body is ready in to a cached byte
array and then wrapped in a new request object. The filter passes this
new wrapped request to the downstream filters on the chain.

TIS21-4067