feat: add DTO validation to issue endpoints
Currently it is possible to send any signed data to any issue endpoint
and it would be accepted. For example sending a placement to the
programme membership endpoint would populate the `tisId`, `startDate`
and `endDate` leaving the `programmeName` as `null`.
The gateway appears to accept
null/empty/missing fields in the credential, so we must validate them
ourselves to avoid any issues. In our example of sending a placement to
the programme membership endpoint it would result in a seemingly valid
Programme Membership credential being issued with no programme name and
other field values from unrelated data.

Add validation to the issue endpoints to avoid empty fields being
accepted.

TIS21-4066
TIS21-4070
TIS21-4071
TIS21-4074