refactor: split the programme membership DTO
Split the programme membership DTO in to "data" for incoming TIS data
and "credential" for outgoing credential data. This will allow additions
to the credential DTO without any risk of modifying what is accepted as
an input via the API.

Create a credential data mapper to convert from the source data to
output credential data.
Update IssueResource to convert the DTO with the mapper before calling
the GatewayService.

Hard code the credential scopes in GatewayServiceTest so that any change
to the hardcoded implementation value will trigger a test failure.

TIS21-4066
TIS21-4141