feat: POC for identity verification
A user should be required to verify their identity, using a DSP
credential, before a TSS credential is able to be issued. The general
flow is designed as below, assuming no previous iteration with the
credential service.

 1. The user requests to issue a TSS credential
 2. The issuance is rejected, with guidance to verify identity
 3. The user sends signed PersonalDetails data to `/api/verify/identity`
 4. The user is sent to the credential gateway to provide an identity
 5. The gateway returns the token code to
    `/api/verify/identity/callback`
 6. We request the full token from the gateway and extract the claims
 7. We compare the claims to the original PersonalDetails, if matching a
    "verified session" is cached.
 8. The user sends a new request to issue a TSS credential
 9. The issuance is allowed as a verified session is available

---

Update application.yml and GatewayProperties to add new properties for
verification endpoints.

Create VerifyResource with endpoints for requesting an identity
credential and then retrieving the associated token. The logic in this
class would normally be held within a `VerificationService` class
instead of directly in the controller.

Create a CacheService to cache the identity data, identity verification
status and code verifiers that need to persist between requests. This
should be replaced with a Redis backed cache, possibly a Redis backed
HttpSession.

Create a VerifiedIdentityFilter which checks whether the user has
verified their identity, requests to `/issue` endpoints will be blocked
without prior verification.

TIS21-4108