feat: create endpoint to start id verification
Create an API endpoint to start the identity verification journey, the
endpoint will accept and cache user identity data and then create a
request URI the user can follow to provide an identity credential to the
gateway.

Create a Redis caching layer, this will be used to cache data that must
be available across requests. Such as identity data, client state and
PKCE code verifiers. Internally generated states and nonces should be
used as cache keys as appropriate.

Create a verification service which will manage the caching of required
data and the generation of the URI to direct users to for verification.

Create an endpoint at `/api/verify/identity` which calls the service to
start the verification flow.

TIS21-4181
TIS21-4182