feat: cache session ID in identity verify start
After verifying an identity a cached session ID should be used to allow
credentials to be issued within the duration of that session.
Update the `/api/verify/identity` to get the auth token from the headers
and pass to `VerificationService.startIdentityVerification()`.
The VerificationService will take care of caching the session
identifier.

Cognito tokens include `origin_jti` in the id/access tokens, which is
consistent within the same session (based on refresh token).
Caching that value will allow any Cognito token within the current sign
in session to be used, eliminating any issues with token refresh.

TIS21-4177
TIS21-4183