feat: allow token verification when getting claims
Create an overloaded `getClaims` method allowing toggling of the
signature verification.
The existing `getClaims` method will not change, so will not verify the
token.

Create a key resolver to get the public key for the token, based on the
issuer and known JWKS endpoints. The retrieved public key will be cached
for later usage to avoid having to request the key each and every time.

TIS21-4177
TIS21-4222