feat: invalidate signatures when data modified
Signed data is currently trusted to always be accurate, with the
signature valid for a set length of time.

Add an additional check to the `SignedDataFilter` to ensure that signed
data is rejected if the underlying data was modified after signing.

Create a RevocationService function to get the last modified data for a
given credential type and TIS ID combination.

TIS21-4128
TIS21-4304