feat: cache timestamp when starting issuance
Cache the start time when beginning the issuance process, the timestamp
will be used later to check for data expiration during the gateway
issuance process.
These checks cover edge cases when the gateway interaction takes an
extended amount of time and the underlying data is modified before the
issuance is complete.

TIS21-4147
TIS21-4346