feat: revoke issued credential if stale
Revoke a credential straight after issuing if the data was modified
since the issuance request began.
Once we send the user to the credential gateway we wait for a callback
to be called, if the underlying data changed during the time the user is
interacting with the gateway it would lead to an invalid credential in
their wallet and the gateway.
Check the last modified time against the issuance start time when the
callback is hit to determine whether the credential should be instantly
revoked.

TIS21-4147
TIS21-4347