fix: use current time for modified timestamp
The last modification timestamp uses the timestamp from the update
message if it is available, however that can lead to incorrect behaviour
depending on the delay between the update/delete occuring at source and
it being received by this service.

Example theoretical timeline:
 - 00:00 - record created in TIS
 - 00:10 - record received by credentials service, last modified date
   stored
 - 00:20 - record deleted in TIS and added to queues
 - 00:30 - user starts issuance process, start time recorded as `00:30`
 - 00:40 - delete event received by credentials service, last modified
   recorded, from source, as `00:20`
 - 00:50 - user completes issuance service, callback compares the
   issuance start time `00:30` against the last modified time `00:20` and
   determines that the data was not stale.

By changing the last modified to always be the current time, the last
modified would be `00:40`. Which when compared to the issuance start
time `00:30` would flag the data as stale correctly.

TIS21-4147
TIS21-4347