feat: include unique identifier in issued creds
Issued credentials must include a unique identifier that matches the
unique identifier from the verified identity credential.

Update the credential DTOs to include a `uniqueIdentifier` field, this
should be used to populate the `UniqueIdentifier` claim during issuance.

Refactor `IssuanceService.startCredentialIssuance` to accept a new
`TisDataDto` instead of `CredentialDto`, as well as a mapper function
allow the service to map from the data DTO to the credential DTO.
The unique identifier will be populated in the credential data by
checking against the cached verified session.

Update the endpoints for issuing PM and placement credentials to provide
the data DTO instead of the pre-mapped credential data. Generate the
mapping function to allow the service to generically handling mapping of
multiple credential types.

Update the credential data mapper to also populate the unique identifier
during conversion.

TIS21-4932
TIS21-4849