fix: tracing filter registration
Update the AWS Xray SDK to `2.14.0`, which allows it to support Spring
6's Jakarta based Filters.
Move the tracing filter configuration to FilterConfiguration as the
tracing filter must be placed before the SignedDataFilter in the filter
chain. Failure to register them in this order will cause the trace to
fail for any unsigned data, as the SignedDataFilter returns an error
instead of continuing the filter chain calls.

TIS21-4226