feat: emit cred event after revocation
When a credential is revoked a credential event should be published to
allow other services, such as notifications, to act accordingly.

Create an EventPublishingService to handle to publish to SNS, the group
ID should be set to the credential's ID to allow better de-duplication
and ensure that we cannot ever get multiple downstream instances
processing events for the same credential simultaneously.

TIS21-5115
TIS21-5133