feat: add fuzzy name matching during id checks
The identity checks rely on matching the forename and surname between
TIS and the provided DSP credential. As there is a high percentage
chance that names will mismatch between spelling mistakes and
inclusive/exclusion of middle name, the name checking needs to be more
forgiving.

Update the VerificationService to check the phonetic and text accuracies
and make a decision based on the similarity rather than exact matching.
Split names on spaces and hypens to allow matching when middle names, or
double barrelled names are used.

Publish metrics for both phonetic and text accuracy to Cloudwatch to
allow monitoring and ongoing improvements to be made to the level of
fuzziness allowed.

TIS21-