refactor: invert the accuracy metric
The micrometer meters do not support publishing of minimum values, only
max and average (of those we care about for accuracy). As a result if
accuracy is reported on then we can only ever see the best matches, but
being able to monitor the worst matches is more useful to us.

Update the meter configuration to rename the metric from
`identity.accuracy` to `identity.inaccuracy`.
Update the MetricsService to invert the accuracy value before
publishing.

TIS21-6065
TIS21-6092